+++ src/components/TasksList.tsx
@@ -1,15 +1,20 @@
 import { Stack } from "@mui/material";
 import TaskItem from "./TaskItem";
 import useTaskStore from "../store/taskStore";
+import useFilterTasks from "../hooks/useFilterTasks";
+
+type props = {
+  filter: string;
+};
 
 // Renders a list of task items
-const TasksList = () => {
-  const tasks = useTaskStore((state) => state.tasks);
+const TasksList = ({ filter }: props) => {
+  const tasks = useFilterTasks(filter);
   return (
     <Stack gap="0.5rem">
       {tasks.map((task) => (
         <TaskItem task={task} />
-      ))}{" "}
+      ))}
     </Stack>
   );
 };