+++ src/utils/functions/getEmptyTask.ts
@@ -1,7 +1,7 @@
 /*** Return an empty, fresh task of type taskType. ***/
 
 import { v4 as uuidv4 } from "uuid";
-import taskType from "../types/taskType";
+import taskType from "../../types/taskType";
 
 const getEmptyTask = (): taskType => {
   return {