+++ changelog.md
@@ -1,5 +1,9 @@
 # Change Log
 
+**19th January 2025**
+
+- Filter tasks by text within their name or description.
+
 **18th January 2025**
 
 - Added some dummy tasks on load for testing purposes.
@@ -8,11 +12,11 @@
 - Priority of tasks depicted as a colour on the left border of a task (no colour means no priority, green means low, orange means medium, red means high).
 - If the description of a task item is too long to fit onto one line, the overflow will be hidden.
 
-(A week break due to university exams)
+_(Not many changes + a week break due to university exams)_
 
 **11th January 2025**
 
-- Can view description of asks.
+- Can view description of tasks.
 
 **10th January 2025**